FISH-8908 - moved e2e tests to root and removed unit tests
