+++ test/e2e/tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"extends": "../../../tsconfig.json",
+	"extends": "../../tsconfig.json",
 	"compilerOptions": {
 		"target": "es2022",
 		"module": "node16",
